feat(psdl_vocab_search): extract modular search engine into package
Fourth module extraction of the three-tier refactor. Moves the
embedders/retrievers/rerankers/factory engine into an installable
psdl_vocab_search package (flat layout, package-dir={'psdl_vocab_search':'.'}).
app/services/vocabulary_search/ is now a shim re-exporting the package so
terminology_anchoring's lazy import keeps working. Per-module shims
(base, embedders, retrievers, rerankers, factory) added so test_rerankers.py
and any direct sub-module imports keep working. Heavy ML libs
(sentence-transformers/torch/faiss) are optional [ml] extras; numpy is the
only hard dep.

Verified: app.main boots from backend cwd; uvicorn 200; full suite green
(no new failures vs 73-passed baseline).
