+++ backend/requirements.txt
@@ -8,3 +8,5 @@ psdl-lang>=0.4.0  # Always track latest psdl-lang
 python-docx>=0.8.11  # Word document generation for IRB export
 pyarrow>=15,<17
 meds>=0.3,<0.4
+# psdl_vocab_search is installed editable from ./psdl_vocab_search during development
+# (pip install -e ./psdl_vocab_search). ML embedders are optional extras: .[ml]
